refactor [server]: rename application to service
