+++ server/src/modules/article/article.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { ArticleService } from './application/article.service';
 import { ArticleController } from './controller/article.controller';
 import { ArticleSchema } from './repository/typeorm/schema/article.schema';
+import { ArticleService } from './service/article.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([ArticleSchema])],
